Hide away internals and comment, comment, comment
