+++ untangled-2018/game/systems/rendersystem.py
@@ -1,34 +1,16 @@
-from ecs.systems.system import System
-from ecs.components.component import *
-
 import pygame
 from pygame import Rect
 
+from lib.system import System
+from game.components import *
+
 class RenderSystem(System):
-    def __init__(self, screen: pygame.Surface):
+    """This system draws any entity with a SpriteSheet component."""
+
+    def __init__(self, screen):
         self.screen = screen
         self.image_cache = {}
         self.steps = 0
-
-    def get_image(self, spritesheet, index):
-        if spritesheet.path not in self.image_cache:
-            # Load from file
-            sheet_img = pygame.image.load(spritesheet.path)
-            print(sheet_img)
-
-            # Check the file can be divided right
-            if sheet_img.get_width() % spritesheet.tile_size != 0 or sheet_img.get_height() % spritesheet.tile_size != 0:
-                raise ValueError('Spritesheet width and height are not a multiple of its tile size')
-            
-            # Partition into sub-images
-            images = []
-            for y in range(0, sheet_img.get_height(), spritesheet.tile_size):
-                for x in range(0, sheet_img.get_width(), spritesheet.tile_size):
-                    bounds = pygame.Rect(x, y, spritesheet.tile_size, spritesheet.tile_size)
-                    images.append(sheet_img.subsurface(bounds))
-            self.image_cache[spritesheet.path] = images
-
-        return self.image_cache[spritesheet.path][index]
 
     def update(self, game, dt: float, events: list):
         # Step through 15 sprite frames each second
@@ -51,6 +33,7 @@ class RenderSystem(System):
             if IngameObject in entity and SpriteSheet in entity:
                 spritesheet = entity[SpriteSheet]
 
+                # Where are they relative to us?
                 pos = entity[IngameObject].position
                 rel_pos = (pos[0] - our_center[0], pos[1] - our_center[1])
                 screen_pos = (rel_pos[0] + 512, rel_pos[1] + 512)
@@ -75,3 +58,24 @@ class RenderSystem(System):
                 
                 # Draw the image
                 self.screen.blit(img, rect)
+
+    def get_image(self, spritesheet, index):
+        # Ideally, we cache so we only process a file once
+        if spritesheet.path not in self.image_cache:
+            # Load from file
+            sheet_img = pygame.image.load(spritesheet.path)
+
+            # Check the file can be divided right
+            if sheet_img.get_width() % spritesheet.tile_size != 0 or sheet_img.get_height() % spritesheet.tile_size != 0:
+                raise ValueError('Spritesheet width and height are not a multiple of its tile size')
+            
+            # Partition into sub-images
+            images = []
+            for y in range(0, sheet_img.get_height(), spritesheet.tile_size):
+                for x in range(0, sheet_img.get_width(), spritesheet.tile_size):
+                    bounds = pygame.Rect(x, y, spritesheet.tile_size, spritesheet.tile_size)
+                    images.append(sheet_img.subsurface(bounds))
+            self.image_cache[spritesheet.path] = images
+
+        return self.image_cache[spritesheet.path][index]
+
