+++ untangled-2018/install.bat
@@ -9,4 +9,4 @@ echo What would you like the hostname to be?
 set /p HOSTNAME=
 @echo %HOSTNAME%
 
-@echo HOSTNAME="%HOSTNAME%"> ecs/config.py
+@echo HOSTNAME="%HOSTNAME%"> lib/config.py
