+++ untangled-2018/lib/system.py
@@ -1,0 +1,10 @@
+class System:
+    """Updates our game's list of entities."""
+
+    def __init__(self):
+        pass
+
+    def update(self, game, dt, events):
+        """Modify entities and their components based on events and their
+        current state."""
+        pass
